requirements.txt
environs~=14.5.0
aiogram~=3.25.0
PyYAML~=6.0.3
asyncpg~=0.31.0
aiohttp~=3.13.3